+++ site/homepage/content/release-notes.md
@@ -10,6 +10,7 @@ description = "Information about changes in recent Hono releases. Includes new f
 
 * Letting the trace sampling settings of the Hono components be defined via sampling strategies served by the Jaeger
   Collector did not work for components using Quarkus native images. This has been fixed.
+* Command messages with no payload could not be sent to an MQTT device. This has been fixed.
 
 ## 1.7.3